- moved prioritized querset func to util to be used in different locations - moved ParserErrorCategoryChoices to util.py to avoid circular import - updated get_status to reject file if more than 500 priority errors are generated.
